fix(hol-guard): dedupe supply chain manager stats and PATH copy
Extract shared manager coverage counting into supply-chain-protection-stats
and use clearer missing-PATH messaging when no shims are installed yet.
